Finaliza projeto funcionando

+++ src/pages/Repositories/index.js
@@ -1,27 +1,33 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './styled';
+import { useNavigate } from 'react-router-dom';
 
 export default function Repositories() {
-    const [ repositories, setRepositories ] = useState([]);
+    const navegate = useNavigate()
+  const [ repositories, setRepositories ] = useState([]);
 
-    useEffect(() => {
-        let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
-    }, []);
+  useEffect(() => {
+    let repositoriesName = localStorage.getItem('repositoriesName');
+    if(repositoriesName !== null) {
+      repositoriesName = JSON.parse(repositoriesName);
+      setRepositories(repositoriesName);
+      localStorage.clear();
+    } else {
+        navegate.push('/')
+    }
+  }, []);
 
-
-    return (
-        <S.Container>
-            <S.Title>Repositórios</S.Title>
-            <S.List>
-                { repositories.map(repository => {
-                    return (
-                        <S.ListItem>Repositório: { repository }</S.ListItem>
-                    )
-                }) }
-            </S.List>
-        </S.Container>
-    )
+  return (
+    <S.Container>
+      <S.Title>Repositórios</S.Title>
+      <S.List>
+        { repositories.map(repository => {
+          return (
+            <S.ListItem>Repositório: { repository }</S.ListItem>
+          )
+        }) }
+      </S.List>
+      <S.LinkHome to="/">Voltar</S.LinkHome>
+    </S.Container>
+  )
 }
